requirements.txt
cached-property==1.5.2
certifi==2020.6.20
chainer==7.8.0
chainerrl==0.8.0
cloudpickle==1.6.0
cupy-cuda100==8.6.0
cycler==0.10.0
fastrlock==0.8
filelock==3.0.4
future==0.18.2
gym==0.17.3
imageio==2.9.0
kiwisolver==1.1.0
matlabengineforpython===R2017a
matplotlib==3.0.3
numpy==1.18.5
olefile==0.46
opencv-python==3.4.1.15
Pillow==7.2.0
protobuf==3.5.2
pyglet==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
scipy==1.4.1
six==1.16.0
typing-extensions==3.10.0.2
